requirements.txt
# Core ML (latest compatible)
torch>=2.3.0,<3.0
torchvision>=0.18.0,<1.0
torchaudio>=2.3.0,<3.0

# Transformers
transformers>=4.42.0,<5.0
accelerate>=0.30.0

# YOLO
ultralytics>=8.2.50

# Computer Vision
opencv-python>=4.10.0
Pillow>=10.4.0

# Depth
timm>=1.0.0

# Scientific
numpy>=1.26.4
# scipy removed — not used anywhere in the codebase

# Backend
fastapi>=0.111.0
uvicorn>=0.30.0
python-multipart>=0.0.9   # required by FastAPI for UploadFile / form-data
websockets>=12.0           # required by uvicorn WebSocket support

# Utils
requests>=2.31.0
tqdm>=4.66.0